Fixed vncExtInit.cc compilation with gcc 4 series (vncExtInit.cc:754: error:cast from 'void*' to 'CARD32' loses precision)
git-svn-id: svn://svn.code.sf.net/p/tigervnc/code/branches/1.5-xserver@2443 3789f03b-4d11-0410-bbf8-ca57d06f2519

+++ unix/xserver/hw/vnc/vncExtInit.cc
@@ -751,7 +751,7 @@ static int ProcVncExtGetQueryConnect(ClientPtr client)
   rep.timeout = qcTimeout;
   rep.addrLen = qcTimeout ? strlen(qcAddress) : 0;
   rep.userLen = qcTimeout ? strlen(qcUsername) : 0;
-  rep.opaqueId = (CARD32)queryConnectId;
+  rep.opaqueId = (CARD32)(long)queryConnectId;
   rep.length = (rep.userLen + rep.addrLen + 3) >> 2;
   if (client->swapped) {
     swaps(&rep.sequenceNumber, n);
